Separate FloatingPointFunctions protocol
